+++ apps/web/components/editor/plugins/toc-kit.tsx
@@ -1,0 +1,8 @@
+'use client'
+
+import { BaseTocPlugin } from '@platejs/toc'
+import { toTPlatePlugin } from 'platejs/react'
+
+import { TocElement } from '@/components/ui/toc-node'
+
+export const TocKit = [toTPlatePlugin(BaseTocPlugin).withComponent(TocElement)]